Nullptr check, tutorial added, Take() bug fixes
Added some more nullptr checks, that even though some might be a bit redundant make the code safer. Added a set of instructions on how to finish the Game. Fixed a few bugs with the Take() command that didn't allow the user to take certain objects and that allowed him/her to take objects that were already taken. Added a wait for Enter when the game is finished, so the user can know he/she has won.